# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "InvoiceTool Swagger API",
        "description": "InvoiceTool Swagger API description",
        "contact": {
            "email": "admin@finpay.com"
        },
        "version": "1.0.0"
    },
    "host": "invoice-tool-nov8.test/",
    "basePath": "api/v1",
    "paths": {
...
        "/PaymentMethod": {
            "get": {
                "tags": [
                    "PAYMENT METHODFORM"
                ],
                "summary": "Paymentmethodform Table",
                "operationId": "pid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "header",
                        "description": "Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/paymentmethod": {
            "post": {
                "tags": [
                    "PAYMENT METHODFORM"
                ],
                "summary": "Paymentmethod Store",
                "description": "Paymentmethodform Store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Paymentmethod Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/Payment": {
            "get": {
                "tags": [
                    "PAYMENT"
                ],
                "summary": "Payment table",
                "operationId": "Peid",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/Payment/{id}": {
            "get": {
                "tags": [
                    "PAYMENT"
                ],
                "summary": "Payment Edit",
                "operationId": "Peid",
                "parameters": [
                    {
                        "name": "payment_id",
                        "in": "path",
                        "description": "Payment Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PAYMENT"
                ],
                "summary": "Payment Update",
                "description": "Payment Update",
                "parameters": [
                    {
                        "name": "payment_id",
                        "in": "path",
                        "description": "Payment Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payment_amount",
                        "in": "formData",
                        "description": "Payment Amount",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "paymentype_id",
                        "in": "formData",
                        "description": "Payment Type",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "payment_date",
                        "in": "formData",
                        "description": "Payment Date",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "note",
                        "in": "formData",
                        "description": "Note",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PAYMENT"
                ],
                "summary": "Payment Delete",
                "operationId": "Pdid",
                "parameters": [
                    {
                        "name": "payment_id",
                        "in": "path",
                        "description": "Payment Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/paymethod": {
            "get": {
                "tags": [
                    "PAYMENT"
                ],
                "summary": "Payment table",
                "operationId": "Pmid",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
}